backend-service/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.1.1
pymongo==4.3.3
kafka-python==2.0.2
pydantic==1.10.13
python-multipart==0.0.6